+++ src/config/datamodel/EdFi.DmsConfigurationService.DataModel/Model/Application/ApplicationInsertCommand.cs
@@ -12,11 +12,11 @@ namespace EdFi.DmsConfigurationService.DataModel.Model.Application;
 public class ApplicationInsertCommand
 {
     public string ApplicationName { get; set; } = "";
-    public long VendorId { get; set; }
+    public int VendorId { get; set; }
     public string ClaimSetName { get; set; } = "";
     public long[] EducationOrganizationIds { get; set; } = [];
-    public long[] DataStoreIds { get; set; } = [];
-    public long[] ProfileIds { get; set; } = [];
+    public int[] DataStoreIds { get; set; } = [];
+    public int[] ProfileIds { get; set; } = [];
 
     public class Validator : AbstractValidator<ApplicationInsertCommand>
     {
